+++ src/components/index.ts
@@ -2,5 +2,5 @@ export {default as Container} from './Container';
 export {default as Portfolio} from './Portfolio/Portfolio';
 export {default as Search} from './Search';
 export {default as StockDetail} from './stockdetail/StockDetail';
-export {default as StockList} from './StockList';
+export {default as StockList} from './StockList/StockList';
 export {default as Trade} from './Trade';
